feat: refactor ApuBudgetDialog y ApuResourceRow para utilizar utilidades de tipo de recurso
